TL:DR of description: Google Maps API stuff.
Reset API Key, verified SHA1 Certificate Fingerprint manually and
compared it to the one registered with Google, and checked that the API
Key Identifier was correct in both the manifest and in
res/values/google_maps_api. Unfortunately, there still appears to be a
verification error. Recall that another group had a functioning app
utilizing what was presumably Google Maps API; perhaps solicit help on
the subject from them?